Module docugen::document

Structs

DocumentTemplate
A DocumentTemplate mimics a mustache template. A template consists of a list of Partials.

FilledDocument
A FilledDocument is generated from a DocumentTemplate with the required Tags filled in.

TagPair
A TagPair is an association between the tag name key and the value that should be used to fill its place.

Enums

Partial
Each Partial is either a UTF-8 StringLiteral, or a Tag.

TemplateError
Cause of error when trying to fill a DocumentTemplate.

Type Definitions

Identifier